chore: restructure projects folder and simplify queries
