+++ apps/web/components/sections/home/hero.tsx
@@ -5,7 +5,7 @@ import { motion } from 'framer-motion'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { Button } from '~/components/base/button'
-import { CategoryBadge } from '~/components/sections/projects/category-badge'
+import { CategoryBadge } from '~/components/sections/projects/filters'
 import { CategoryBadgeSkeleton } from '~/components/sections/projects/skeletons'
 import { getAllCategories } from '~/lib/queries/projects'
