+++ apps/web/components/sections/projects/cards/index.ts
@@ -1,0 +1,2 @@
+export * from './project-card-grid'
+export * from './project-card-list'
