+++ apps/web/components/sections/projects/detail/animated-counter.tsx
