+++ apps/web/components/sections/projects/detail/file-icon.tsx
