+++ apps/web/components/sections/projects/detail/project-client-wrapper.tsx
