+++ apps/web/components/sections/projects/detail/project-tabs.tsx
