+++ apps/web/components/sections/projects/detail/skeletons/breadcrumb-skeleton.tsx
