+++ apps/web/components/sections/projects/detail/skeletons/index.ts
