+++ apps/web/components/sections/projects/detail/skeletons/project-hero-skeleton.tsx
