+++ apps/web/components/sections/projects/detail/skeletons/project-sidebar-skeleton.tsx
