+++ apps/web/components/sections/projects/detail/skeletons/project-tabs-skeleton.tsx
