+++ apps/web/components/sections/projects/detail/tabs/index.ts
