+++ apps/web/components/sections/projects/filters/category-filters.tsx
@@ -6,7 +6,7 @@ import { X } from 'lucide-react'
 import type { Tables } from '@services/supabase'
 import { Button } from '~/components/base/button'
 import { fadeIn, staggerContainer } from '~/lib/constants/animations'
-import { CategoryBadge } from './category-badge'
+import { CategoryBadge } from './'
 
 interface CategoryFiltersProps {
 	categories: Tables<'categories'>[]
