+++ apps/web/components/sections/projects/filters/index.ts
@@ -1,0 +1,4 @@
+export * from './category-badge'
+export * from './category-filters'
+export * from './sort-dropdown'
+export * from './view-toggle'
