+++ apps/web/lib/queries/projects/index.ts
@@ -1,0 +1,8 @@
+export * from './get-all-categories'
+export * from './get-all-projects'
+export * from './get-project-by-id'
+export * from './get-project-milestones'
+export * from './get-project-pitch'
+export * from './get-project-questions'
+export * from './get-project-team'
+export * from './get-project-updates'
